Resolve "Bug fix and refactor"

Closes #21

See merge request courses/dit342/2024/group-01-web!24

+++ server/utils/imageHandler.js
@@ -6,12 +6,11 @@ const { except, removeFromArray } = require("./utils");
 
 /**
  * Adds images to the given post and updates the file usage in the DB.
- * @param item Post/comment to update. Must have an 'images' field!
+ * @param images Images list to update.
  * @param files List of new files
- * @returns Array of hashes in use after update
  */
-async function updateImages(item, files) {
-    const prevHashes = item.images || [];
+async function updateImages(images, files) {
+    const prevHashes = images || [];
     var newHashes = [];
 
     for (const file of files) {        
@@ -24,13 +23,24 @@ async function updateImages(item, files) {
 
     toAdd.forEach(hash => {
         addUsage(hash);
-        item.images.push(hash);
+        images.push(hash);
     });
 
     toRemove.forEach(hash => {
         removeUsage(hash);
-        removeFromArray(item.images, hash);
+        removeFromArray(images, hash);
     });
+}
+
+async function changeImage(image, newFile) {
+    if (image) 
+        removeUsage(image);
+    const hash = await saveFile(newFile);
+    if (hash == image) 
+        return hash;
+
+    await addUsage(hash);
+    return hash;
 }
 
 async function addUsage(hash) {
@@ -39,7 +49,7 @@ async function addUsage(hash) {
 
 async function removeUsage(hash) {
     const image = await models.Images.findOneAndUpdate({hash : hash}, {$inc: {usageCount: -1}}, {new: true});
-    if (image.usageCount <= 0) {
+    if (image && image?.usageCount <= 0) {
         fs.rmSync(path.join(appRoot, "/uploads/" + image.hash), { recursive: true, force: true }); // delete the directory
         await image.deleteOne().exec(); // delete the DB entry
     }
@@ -67,4 +77,4 @@ async function saveFile(file) {
     return imageEntry.hash; // insert new entry into Images collection
 }
 
-module.exports = {updateImages, addUsage, removeUsage}
+module.exports = {updateImages, changeImage, addUsage, removeUsage}
